Clean fix for UI bug (see Description)
Person commands were not specifying an explicit family ID, so the 'aggregateId' was null/empty, which was causing these commands to never correctly update the visibleAggregatesData. It's more correct to update the visible aggregates that match the aggregate(s) being returned by the server, which also solves the UI bug since the correct family will be updated automatically.